+++ src/getPps.cc
@@ -1,22 +1,24 @@
+#include <stddef.h>
+
 #include <Rcpp.h>
 
 // [[Rcpp::export(.getpps_cpp)]]
 Rcpp::NumericVector getpps_cpp(
   Rcpp::NumericVector &x,
-  int n
+  size_t n
 ) {
   // Assume < 0 and equality has already been checked
-  int N = x.length();
+  size_t N = x.length();
   Rcpp::NumericVector prob(N);
-  int *index = new int[N];
-  int one = 0;
-  int oneplus = 0;
-  int oneminus = 0;
+  size_t* index = new size_t[N];
+  size_t one = 0;
+  size_t oneplus = 0;
+  size_t oneminus = 0;
 
 
 
   double xsum = 0.0;
-  for (int i = 0; i < N; i++) {
+  for (size_t i = 0; i < N; i++) {
     if (x[i] < 0.0)
       std::range_error("elements in x must be >= 0.0");
 
@@ -24,7 +26,7 @@ Rcpp::NumericVector getpps_cpp(
   }
 
   double temp = ((double) n) / xsum;
-  for (int i = 0; i < N; i++) {
+  for (size_t i = 0; i < N; i++) {
     prob[i] = x[i] * temp;
 
     if (prob[i] == 1.0) {
@@ -43,21 +45,21 @@ Rcpp::NumericVector getpps_cpp(
     return prob;
   }
 
-  int onesum = one + oneplus;
+  size_t onesum = one + oneplus;
   while (oneplus > 0) {
     xsum = 0.0;
-    for (int i = 0; i < oneminus; i++)
+    for (size_t i = 0; i < oneminus; i++)
       xsum += x[index[i]];
 
     temp = ((double) (n - onesum)) / xsum;
 
-    int toneminus = oneminus;
+    size_t toneminus = oneminus;
     one = 0;
     oneplus = 0;
     oneminus = 0;
 
 
-    for (int i = 0; i < toneminus; i++) {
+    for (size_t i = 0; i < toneminus; i++) {
       prob[index[i]] = x[index[i]] * temp;
 
       if (prob[index[i]] == 1.0) {
